+++ apps/web/src/lib/db/schema/modules/notifications.ts
@@ -11,10 +11,10 @@ export async function createTables<T>(db: Kysely<T>): Promise<void> {
     .addColumn("phone", "varchar(9)")
     .addColumn("phone_verified", "integer", (col) => col.notNull().defaultTo(0))
     .addColumn("phone_verification_code", "text")
-    .addColumn("phone_verification_expires_at", "integer")
+    .addColumn("phone_verification_expires_at", "bigint")
     .addColumn("notify_on_completion", "integer", (col) => col.notNull().defaultTo(1))
     .addColumn("notify_on_failure", "integer", (col) => col.notNull().defaultTo(1))
-    .addColumn("created_at", "integer", (col) => col.notNull())
-    .addColumn("updated_at", "integer", (col) => col.notNull())
+    .addColumn("created_at", "bigint", (col) => col.notNull())
+    .addColumn("updated_at", "bigint", (col) => col.notNull())
     .execute();
 }
